Moved PullShubImage to singularity package
